+++ app.py
@@ -34,6 +34,7 @@ from fast_api.routes.labeling import router as labeling_router
 from fast_api.routes.record_ide import router as record_ide_router
 from fast_api.routes.record import router as record_router
 from fast_api.routes.weak_supervision import router as weak_supervision_router
+from fast_api.routes.labeling_tasks import router as labeling_tasks_router
 from middleware.database_session import DatabaseSessionHandler
 from starlette.applications import Starlette
 from starlette.graphql import GraphQLApp
@@ -61,6 +62,7 @@ from route_prefix import (
     PREFIX_RECORD_IDE,
     PREFIX_RECORD,
     PREFIX_WEAK_SUPERVISION,
+    PREFIX_LABELING_TASKS,
 )
 from util import security, clean_up
 
@@ -109,6 +111,9 @@ fastapi_app.include_router(
 fastapi_app.include_router(record_router, prefix=PREFIX_RECORD, tags=["record"]),
 fastapi_app.include_router(
     weak_supervision_router, prefix=PREFIX_WEAK_SUPERVISION, tags=["weak-supervision"]
+)
+fastapi_app.include_router(
+    labeling_tasks_router, prefix=PREFIX_LABELING_TASKS, tags=["labeling-tasks"]
 )
 
 routes = [
